First draft of KernelAbstractions
KernelAbstractions.jl is the successor to GPUifyLoops.jl

The idea is that users can write portable code that is performance
tuned for the GPU and that does the right thing by default.

KernelAbstractions.jl provides a small kernel language that can be
used inside the `@kernel` macro to access and write kernels, which
are executable on both the CPU (threaded and non-threaded) as well
as Nvidia GPUs. It should be relatively straightforward to port it
to AMD and Intel GPUs in the future.

To a brave new world!